+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v0.7.0
+- Added support for querying IS NULL and IS NOT NULL in the query builder comparator.
+- Added a method to retrieve the count from a query builder.
+
 ## v0.6.0
 - Added support to multiple order by in query.